+++ vpplink/binapi/vppapi/af_xdp/af_xdp.ba.go
@@ -178,7 +178,7 @@ func (*AfXdpCreateReply) GetMessageType() api.MessageType {
 	return api.ReplyMessage
 }
 func (m *AfXdpCreateReply) GetRetVal() error {
-	return api.RetvalToVPPApiError(m.Retval)
+	return api.RetvalToVPPApiError(int32(m.Retval))
 }
 
 func (m *AfXdpCreateReply) Size() (size int) {
@@ -286,7 +286,7 @@ func (*AfXdpCreateV2Reply) GetMessageType() api.MessageType {
 	return api.ReplyMessage
 }
 func (m *AfXdpCreateV2Reply) GetRetVal() error {
-	return api.RetvalToVPPApiError(m.Retval)
+	return api.RetvalToVPPApiError(int32(m.Retval))
 }
 
 func (m *AfXdpCreateV2Reply) Size() (size int) {
@@ -361,7 +361,7 @@ func (*AfXdpDeleteReply) GetMessageType() api.MessageType {
 	return api.ReplyMessage
 }
 func (m *AfXdpDeleteReply) GetRetVal() error {
-	return api.RetvalToVPPApiError(m.Retval)
+	return api.RetvalToVPPApiError(int32(m.Retval))
 }
 
 func (m *AfXdpDeleteReply) Size() (size int) {
