+++ vpplink/binapi/vppapi/crypto_sw_scheduler/crypto_sw_scheduler.ba.go
@@ -80,7 +80,7 @@ func (*CryptoSwSchedulerSetWorkerReply) GetMessageType() api.MessageType {
 	return api.ReplyMessage
 }
 func (m *CryptoSwSchedulerSetWorkerReply) GetRetVal() error {
-	return api.RetvalToVPPApiError(m.Retval)
+	return api.RetvalToVPPApiError(int32(m.Retval))
 }
 
 func (m *CryptoSwSchedulerSetWorkerReply) Size() (size int) {
